Do the backend validations && Use permissions validate users with roles and permissions

+++ WebSecService/app/Http/Controllers/UserController.php
@@ -6,25 +6,22 @@ use App\Models\User;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\Rule;
 use Illuminate\Validation\Rules\Password;
+use App\Http\Controllers\Controller;
+use Artisan;
+use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
     use ValidatesRequests;
-    // Login function
+
+    // Authentication methods
     public function login()
     {
         return view('auth.login');
     }
 
-    public function profile(Request $request, User $user)
-    {
-        return view('users.profile', compact('user'));
-    }
-
-    // Do Login  function
     public function doLogin(Request $request)
     {
         if (!Auth::attempt(['email' => $request->email, 'password' => $request->password]))
@@ -36,18 +33,20 @@ class UserController extends Controller
         return redirect()->route('home');
     }
 
-    //  register function
+    public function logout()
+    {
+        Auth::logout();
+        return redirect()->route('home');
+    }
+
     public function register()
     {
         return view('auth.register');
     }
 
-    //  register function
     public function doRegister(Request $request, User $user)
     {
-
         try {
-
             $request->validate([
                 'name' => 'required|string|max:255',
                 'email' => 'required|string|email|max:255|unique:users',
@@ -57,73 +56,90 @@ class UserController extends Controller
             return redirect()->back()->withErrors('Invalid registration information.');
         }
 
-
         $user = new User();
         $user->name = $request->name;
         $user->email = $request->email;
         $user->password = bcrypt($request->password);
         $user->save();
 
-
         Auth::login($user);
-        
+
         return redirect()->route('home');
     }
 
-    //  logout function
-    public function logout()
+    // User profile methods
+    public function profile(Request $request, User $user)
     {
-        Auth::logout();
-        return redirect()->route('home');
+        return view('users.profile', compact('user'));
     }
 
-    /**
-     * Display a listing of the resource.
-     */
+    public function edit(User $user)
+    {
+        if (auth()->id() !== $user->id) {
+            if (!auth()->user()->hasPermissionTo('edit_users')) abort(401);
+        }
+        return view('users.edit', compact('user'));
+    }
+
+    public function edit_pass(User $user)
+    {
+        if (auth()->id() !== $user->id) {
+            if (!auth()->user()->hasPermissionTo('edit_users')) abort(401);
+        }
+        return view('users.change_pass', compact('user'));
+    }
+
+    public function savePass(Request $request, User $user)
+    {
+        if (auth()->id() == $user->id) {
+
+            $request->validate([
+                'password' => ['required', 'confirmed', Password::min(8)->numbers()->letters()->mixedCase()->symbols()]
+            ]);
+
+            if (Hash::check($request->old_password, $user->password)) {
+                if ($request->old_password === $request->password) {
+                    return back()->withErrors('New password cannot be the same as the old password.');
+                }
+                $user->password = bcrypt($request->password);
+                $user->save();
+                return redirect()->route('users.profile', ['user' => $user->id]);
+            } else {
+                return back()->withErrors('Old password is incorrect.');
+            }
+        } elseif (!auth()->user()->hasPermissionTo('edit_users')) abort(401);
+
+        $user->password = bcrypt($request->password); //Secure
+        $user->save();
+        return redirect()->route('users.index');
+    }
+
+    // Resource methods
     public function index()
     {
+        if (!auth()->user()->hasPermissionTo('show_users')) abort(401);
         $users = User::all();
         return view('users.index', compact('users'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create()
     {
+        if (!auth()->user()->hasPermissionTo('add_users')) abort(401);
         $user = new User();
         return view('users.create', compact('user'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(User $user)
-    {
-        return view('users.edit', compact('user'));
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit_pass(User $user)
-    {
-        return view('users.change_pass', compact('user'));
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request, User $user)
     {
+
+        if (!auth()->user()->hasPermissionTo('add_users')) abort(401);
         $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|string|email|max:255|unique:users',
-            'password' => 'required|string|confirmed',
+            'password' => 'required|confirmed',
         ]);
 
         $user = new User();
-
         $user->name = $request->name;
         $user->email = $request->email;
         $user->password = bcrypt($request->password); // Ensure password is always set
@@ -132,17 +148,18 @@ class UserController extends Controller
         return redirect()->route('users.index');
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, User $user)
     {
+        // dd($request->all());
+        if (auth()->id() !== $user->id) {
+            if (!auth()->user()->hasPermissionTo('edit_users')) abort(401);
+        }
         $request->validate([
             'name' => 'required|string|max:255',
             'email' => [
                 'required',
                 'string',
-                'email',    
+                'email',
                 'max:255',
                 Rule::unique('users')->ignore($user->id),
             ],
@@ -152,38 +169,15 @@ class UserController extends Controller
         $user->email = $request->email;
         $user->save();
 
-        return redirect()->route('users.profile');
-    }
-
-    
-    /**
-     * Change the password of the specified user.
-     */
-    public function savePass(Request $request, User $user)
-    {
-        $request->validate([
-            'old_password' => 'required|string',
-            'password' => 'required|string|confirmed|different:old_password',
-        ]);
-
-        if (Hash::check($request->old_password, $user->password)) {
-            if ($request->old_password === $request->password) {
-                return back()->withErrors('New password cannot be the same as the old password.');
-            }
-            $user->password = bcrypt($request->password);
-            $user->save();
-            return redirect()->route('users.profile', ['user' => $user->id]);
-        } else {
-            return back()->withErrors('Old password is incorrect.');
+        if (auth()->id() !== $user->id) {
+            return redirect()->route('users.index');
         }
-
+        return redirect()->route('users.profile', compact('user'));
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(User $user)
     {
+        if (!auth()->user()->hasPermissionTo('delete_users')) abort(401);
         $user->delete();
         return redirect()->route('users.index');
     }
